Continuing refactor
